+++ bevy_rapier2d/examples/testbed2.rs
@@ -205,7 +205,7 @@ fn main() {
             (
                 cleanup,
                 |mut rapier_config: Query<&mut RapierConfiguration>,
-                 ctxt: DefaultRapierContextAccess| {
+                 ctxt: ReadDefaultRapierContext| {
                     let mut rapier_config = rapier_config.single_mut();
                     rapier_config.gravity =
                         RapierConfiguration::new(ctxt.integration_parameters.length_unit).gravity;